+++ src/coprocessor/dag/rpn_expr/types/expr_builder.rs
@@ -1,0 +1,698 @@
+// Copyright 2019 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#![allow(dead_code)]
+
+use std::convert::{TryFrom, TryInto};
+
+use cop_datatype::{EvalType, FieldTypeAccessor};
+use tipb::expression::{Expr, ExprType};
+
+use super::super::function::RpnFunction;
+use super::expr::{RpnExpression, RpnExpressionNode};
+use crate::coprocessor::codec::data_type::ScalarValue;
+use crate::coprocessor::codec::mysql::Tz;
+use crate::coprocessor::codec::mysql::{Decimal, Duration, Json, Time, MAX_FSP};
+use crate::coprocessor::{Error, Result};
+use crate::util::codec::number;
+
+/// Helper to build an `RpnExpression`.
+///
+/// TODO: Deprecate it in Coprocessor V2 DAG interface.
+pub struct RpnExpressionBuilder;
+
+impl RpnExpressionBuilder {
+    /// Builds the RPN expression node list from an expression definition tree.
+    pub fn build_from_expr_tree(tree_node: Expr, time_zone: &Tz) -> Result<RpnExpression> {
+        let mut expr_nodes = Vec::new();
+        append_rpn_nodes_recursively(
+            tree_node,
+            &mut expr_nodes,
+            time_zone,
+            super::super::map_pb_sig_to_rpn_func,
+        )?;
+        Ok(RpnExpression::from(expr_nodes))
+    }
+}
+
+/// Transforms eval tree nodes into RPN nodes.
+///
+/// Suppose that we have a function call:
+///
+/// ```ignore
+/// A(B, C(E, F, G), D)
+/// ```
+///
+/// The eval tree looks like:
+///
+/// ```ignore
+///           +---+
+///           | A |
+///           +---+
+///             |
+///   +-------------------+
+///   |         |         |
+/// +---+     +---+     +---+
+/// | B |     | C |     | D |
+/// +---+     +---+     +---+
+///             |
+///      +-------------+
+///      |      |      |
+///    +---+  +---+  +---+
+///    | E |  | F |  | G |
+///    +---+  +---+  +---+
+/// ```
+///
+/// We need to transform the tree into RPN nodes:
+///
+/// ```ignore
+/// B E F G C D A
+/// ```
+///
+/// The transform process is very much like a post-order traversal. This function does it
+/// recursively.
+fn append_rpn_nodes_recursively<F>(
+    tree_node: Expr,
+    rpn_nodes: &mut Vec<RpnExpressionNode>,
+    time_zone: &Tz,
+    fn_mapper: F,
+) -> Result<()>
+where
+    F: Fn(tipb::expression::ScalarFuncSig) -> Result<Box<dyn RpnFunction>> + Copy,
+{
+    // TODO: We should check whether node types match the function signature. Otherwise there
+    // will be panics when the expression is evaluated.
+
+    match tree_node.get_tp() {
+        ExprType::Null => handle_node_constant_null(tree_node, rpn_nodes),
+        ExprType::Int64 => handle_node_constant_int64(tree_node, rpn_nodes),
+        ExprType::Uint64 => handle_node_constant_uint64(tree_node, rpn_nodes),
+        ExprType::String | ExprType::Bytes => handle_node_constant_bytes(tree_node, rpn_nodes),
+        ExprType::Float32 | ExprType::Float64 => handle_node_constant_float(tree_node, rpn_nodes),
+        ExprType::MysqlTime => handle_node_constant_time(tree_node, rpn_nodes, time_zone),
+        ExprType::MysqlDuration => handle_node_constant_duration(tree_node, rpn_nodes),
+        ExprType::MysqlDecimal => handle_node_constant_decimal(tree_node, rpn_nodes),
+        ExprType::MysqlJson => handle_node_constant_json(tree_node, rpn_nodes),
+        ExprType::ColumnRef => handle_node_column_ref(tree_node, rpn_nodes),
+        ExprType::ScalarFunc => handle_node_fn_call(tree_node, rpn_nodes, time_zone, fn_mapper),
+        t => Err(box_err!("Unexpected ExprType {:?}", t)),
+    }
+}
+
+/// TODO: Remove this helper function when we use Failure which can simplify the code.
+#[inline]
+fn get_eval_type(tree_node: &Expr) -> Result<EvalType> {
+    EvalType::try_from(tree_node.get_field_type().tp()).map_err(|e| Error::Other(box_err!(e)))
+}
+
+#[inline]
+fn handle_node_column_ref(
+    mut tree_node: Expr,
+    rpn_nodes: &mut Vec<RpnExpressionNode>,
+) -> Result<()> {
+    let offset = number::decode_i64(&mut tree_node.get_val()).map_err(|_| {
+        Error::Other(box_err!(
+            "Unable to decode column reference offset from the request"
+        ))
+    })? as usize;
+    rpn_nodes.push(RpnExpressionNode::ColumnRef {
+        offset,
+        field_type: tree_node.take_field_type(),
+    });
+    Ok(())
+}
+
+#[inline]
+fn handle_node_fn_call<F>(
+    mut tree_node: Expr,
+    rpn_nodes: &mut Vec<RpnExpressionNode>,
+    time_zone: &Tz,
+    fn_mapper: F,
+) -> Result<()>
+where
+    F: Fn(tipb::expression::ScalarFuncSig) -> Result<Box<dyn RpnFunction>> + Copy,
+{
+    // Map pb func to `RpnFunction`.
+    let func = fn_mapper(tree_node.get_sig())?;
+    let args = tree_node.take_children().into_vec();
+    if func.args_len() != args.len() {
+        return Err(box_err!(
+            "Unexpected arguments, expect {}, received {}",
+            func.args_len(),
+            args.len()
+        ));
+    }
+    for arg in args {
+        append_rpn_nodes_recursively(arg, rpn_nodes, time_zone, fn_mapper)?;
+    }
+    rpn_nodes.push(RpnExpressionNode::FnCall {
+        func,
+        field_type: tree_node.take_field_type(),
+    });
+    Ok(())
+}
+
+#[inline]
+fn handle_node_constant_null(
+    mut tree_node: Expr,
+    rpn_nodes: &mut Vec<RpnExpressionNode>,
+) -> Result<()> {
+    let eval_type = get_eval_type(&tree_node)?;
+
+    let scalar_value = match eval_type {
+        EvalType::Int => ScalarValue::Int(None),
+        EvalType::Real => ScalarValue::Real(None),
+        EvalType::Decimal => ScalarValue::Decimal(None),
+        EvalType::Bytes => ScalarValue::Bytes(None),
+        EvalType::DateTime => ScalarValue::DateTime(None),
+        EvalType::Duration => ScalarValue::Duration(None),
+        EvalType::Json => ScalarValue::Json(None),
+    };
+    rpn_nodes.push(RpnExpressionNode::Constant {
+        value: scalar_value,
+        field_type: tree_node.take_field_type(),
+    });
+    Ok(())
+}
+
+#[inline]
+fn handle_node_constant_int64(
+    mut tree_node: Expr,
+    rpn_nodes: &mut Vec<RpnExpressionNode>,
+) -> Result<()> {
+    let eval_type = get_eval_type(&tree_node)?;
+
+    let scalar_value = match eval_type {
+        EvalType::Int => {
+            let value = number::decode_i64(&mut tree_node.get_val()).map_err(|_| {
+                Error::Other(box_err!(
+                    "Unable to decode {:?} from the request",
+                    eval_type
+                ))
+            })?;
+            ScalarValue::Int(Some(value))
+        }
+        t => {
+            return Err(box_err!(
+                "Unexpected eval type {:?} for ExprType {:?}",
+                t,
+                tree_node.get_tp()
+            ));
+        }
+    };
+    rpn_nodes.push(RpnExpressionNode::Constant {
+        value: scalar_value,
+        field_type: tree_node.take_field_type(),
+    });
+    Ok(())
+}
+
+#[inline]
+fn handle_node_constant_uint64(
+    mut tree_node: Expr,
+    rpn_nodes: &mut Vec<RpnExpressionNode>,
+) -> Result<()> {
+    let eval_type = get_eval_type(&tree_node)?;
+
+    let scalar_value = match eval_type {
+        EvalType::Int => {
+            let value = number::decode_u64(&mut tree_node.get_val()).map_err(|_| {
+                Error::Other(box_err!(
+                    "Unable to decode {:?} from the request",
+                    eval_type
+                ))
+            })?;
+            ScalarValue::Int(Some(value as i64))
+        }
+        t => {
+            return Err(box_err!(
+                "Unexpected eval type {:?} for ExprType {:?}",
+                t,
+                tree_node.get_tp()
+            ));
+        }
+    };
+    rpn_nodes.push(RpnExpressionNode::Constant {
+        value: scalar_value,
+        field_type: tree_node.take_field_type(),
+    });
+    Ok(())
+}
+
+#[inline]
+fn handle_node_constant_bytes(
+    mut tree_node: Expr,
+    rpn_nodes: &mut Vec<RpnExpressionNode>,
+) -> Result<()> {
+    let eval_type = get_eval_type(&tree_node)?;
+
+    let scalar_value = match eval_type {
+        EvalType::Bytes => ScalarValue::Bytes(Some(tree_node.take_val())),
+        t => {
+            return Err(box_err!(
+                "Unexpected eval type {:?} for ExprType {:?}",
+                t,
+                tree_node.get_tp()
+            ));
+        }
+    };
+    rpn_nodes.push(RpnExpressionNode::Constant {
+        value: scalar_value,
+        field_type: tree_node.take_field_type(),
+    });
+    Ok(())
+}
+
+#[inline]
+fn handle_node_constant_float(
+    mut tree_node: Expr,
+    rpn_nodes: &mut Vec<RpnExpressionNode>,
+) -> Result<()> {
+    let eval_type = get_eval_type(&tree_node)?;
+
+    let scalar_value = match eval_type {
+        EvalType::Real => {
+            let value = number::decode_f64(&mut tree_node.get_val()).map_err(|_| {
+                Error::Other(box_err!(
+                    "Unable to decode {:?} from the request",
+                    eval_type
+                ))
+            })?;
+            ScalarValue::Real(Some(value))
+        }
+        t => {
+            return Err(box_err!(
+                "Unexpected eval type {:?} for ExprType {:?}",
+                t,
+                tree_node.get_tp()
+            ));
+        }
+    };
+    rpn_nodes.push(RpnExpressionNode::Constant {
+        value: scalar_value,
+        field_type: tree_node.take_field_type(),
+    });
+    Ok(())
+}
+
+#[inline]
+fn handle_node_constant_time(
+    mut tree_node: Expr,
+    rpn_nodes: &mut Vec<RpnExpressionNode>,
+    time_zone: &Tz,
+) -> Result<()> {
+    let eval_type = get_eval_type(&tree_node)?;
+    let field_type = tree_node.take_field_type();
+
+    let scalar_value = match eval_type {
+        EvalType::DateTime => {
+            let v = number::decode_u64(&mut tree_node.get_val()).map_err(|_| {
+                Error::Other(box_err!(
+                    "Unable to decode {:?} from the request",
+                    eval_type
+                ))
+            })?;
+            let fsp = field_type.decimal() as i8;
+            let value = Time::from_packed_u64(v, field_type.tp().try_into()?, fsp, time_zone)
+                .map_err(|_| {
+                    Error::Other(box_err!(
+                        "Unable to decode {:?} from the request",
+                        eval_type
+                    ))
+                })?;;
+            ScalarValue::DateTime(Some(value))
+        }
+        t => {
+            return Err(box_err!(
+                "Unexpected eval type {:?} for ExprType {:?}",
+                t,
+                tree_node.get_tp()
+            ));
+        }
+    };
+    rpn_nodes.push(RpnExpressionNode::Constant {
+        value: scalar_value,
+        field_type,
+    });
+    Ok(())
+}
+
+#[inline]
+fn handle_node_constant_duration(
+    mut tree_node: Expr,
+    rpn_nodes: &mut Vec<RpnExpressionNode>,
+) -> Result<()> {
+    let eval_type = get_eval_type(&tree_node)?;
+
+    let scalar_value = match eval_type {
+        EvalType::Duration => {
+            let n = number::decode_i64(&mut tree_node.get_val()).map_err(|_| {
+                Error::Other(box_err!(
+                    "Unable to decode {:?} from the request",
+                    eval_type
+                ))
+            })?;
+            let value = Duration::from_nanos(n, MAX_FSP).map_err(|_| {
+                Error::Other(box_err!(
+                    "Unable to decode {:?} from the request",
+                    eval_type
+                ))
+            })?;
+            ScalarValue::Duration(Some(value))
+        }
+        t => {
+            return Err(box_err!(
+                "Unexpected eval type {:?} for ExprType {:?}",
+                t,
+                tree_node.get_tp()
+            ));
+        }
+    };
+    rpn_nodes.push(RpnExpressionNode::Constant {
+        value: scalar_value,
+        field_type: tree_node.take_field_type(),
+    });
+    Ok(())
+}
+
+#[inline]
+fn handle_node_constant_decimal(
+    mut tree_node: Expr,
+    rpn_nodes: &mut Vec<RpnExpressionNode>,
+) -> Result<()> {
+    let eval_type = get_eval_type(&tree_node)?;
+
+    let scalar_value = match eval_type {
+        EvalType::Decimal => {
+            let value = Decimal::decode(&mut tree_node.get_val()).map_err(|_| {
+                Error::Other(box_err!(
+                    "Unable to decode {:?} from the request",
+                    eval_type
+                ))
+            })?;
+            ScalarValue::Decimal(Some(value))
+        }
+        t => {
+            return Err(box_err!(
+                "Unexpected eval type {:?} for ExprType {:?}",
+                t,
+                tree_node.get_tp()
+            ));
+        }
+    };
+    rpn_nodes.push(RpnExpressionNode::Constant {
+        value: scalar_value,
+        field_type: tree_node.take_field_type(),
+    });
+    Ok(())
+}
+
+#[inline]
+fn handle_node_constant_json(
+    mut tree_node: Expr,
+    rpn_nodes: &mut Vec<RpnExpressionNode>,
+) -> Result<()> {
+    let eval_type = get_eval_type(&tree_node)?;
+
+    let scalar_value = match eval_type {
+        EvalType::Json => {
+            let value = Json::decode(&mut tree_node.get_val()).map_err(|_| {
+                Error::Other(box_err!(
+                    "Unable to decode {:?} from the request",
+                    eval_type
+                ))
+            })?;
+            ScalarValue::Json(Some(value))
+        }
+        t => {
+            return Err(box_err!(
+                "Unexpected eval type {:?} for ExprType {:?}",
+                t,
+                tree_node.get_tp()
+            ));
+        }
+    };
+    rpn_nodes.push(RpnExpressionNode::Constant {
+        value: scalar_value,
+        field_type: tree_node.take_field_type(),
+    });
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::super::expr::tests::*;
+    use super::*;
+
+    use cop_datatype::FieldTypeTp;
+    use tipb::expression::ScalarFuncSig;
+
+    /// For testing `append_rpn_nodes_recursively`. It accepts protobuf function sig enum, which
+    /// cannot be modified by us in tests to support FnA ~ FnD. So let's just hard code some
+    /// substitute.
+    fn fn_mapper(value: ScalarFuncSig) -> Result<Box<dyn RpnFunction>> {
+        // FnA: CastIntAsInt
+        // FnB: CastIntAsReal
+        // FnC: CastIntAsString
+        // FnD: CastIntAsDecimal
+        match value {
+            ScalarFuncSig::CastIntAsInt => Ok(Box::new(FnA)),
+            ScalarFuncSig::CastIntAsReal => Ok(Box::new(FnB)),
+            ScalarFuncSig::CastIntAsString => Ok(Box::new(FnC)),
+            ScalarFuncSig::CastIntAsDecimal => Ok(Box::new(FnD)),
+            _ => unreachable!(),
+        }
+    }
+
+    #[test]
+    #[allow(clippy::float_cmp)]
+    fn test_append_rpn_nodes_recursively() {
+        use crate::util::codec::number::NumberEncoder;
+
+        // Input:
+        // FnD(a, FnA(FnC(b, c, d)), FnA(FnB(e, f))
+        //
+        // Tree:
+        //           FnD
+        // +----------+----------+
+        // a         FnA        FnA
+        //            |          |
+        //           FnC        FnB
+        //        +---+---+      +---+
+        //        b   c   d      e   f
+        //
+        // RPN:
+        // a b c d FnC FnA e f FnB FnA FnD
+
+        let node_fn_a_1 = {
+            // node b
+            let mut node_b = Expr::new();
+            node_b.set_tp(ExprType::Int64);
+            node_b
+                .mut_field_type()
+                .as_mut_accessor()
+                .set_tp(FieldTypeTp::LongLong);
+            node_b.mut_val().encode_i64(7).unwrap();
+
+            // node c
+            let mut node_c = Expr::new();
+            node_c.set_tp(ExprType::Int64);
+            node_c
+                .mut_field_type()
+                .as_mut_accessor()
+                .set_tp(FieldTypeTp::LongLong);
+            node_c.mut_val().encode_i64(3).unwrap();
+
+            // node d
+            let mut node_d = Expr::new();
+            node_d.set_tp(ExprType::Int64);
+            node_d
+                .mut_field_type()
+                .as_mut_accessor()
+                .set_tp(FieldTypeTp::LongLong);
+            node_d.mut_val().encode_i64(11).unwrap();
+
+            // FnC
+            let mut node_fn_c = Expr::new();
+            node_fn_c.set_tp(ExprType::ScalarFunc);
+            node_fn_c.set_sig(ScalarFuncSig::CastIntAsString);
+            node_fn_c
+                .mut_field_type()
+                .as_mut_accessor()
+                .set_tp(FieldTypeTp::LongLong);
+            node_fn_c.mut_children().push(node_b);
+            node_fn_c.mut_children().push(node_c);
+            node_fn_c.mut_children().push(node_d);
+
+            // FnA
+            let mut node_fn_a = Expr::new();
+            node_fn_a.set_tp(ExprType::ScalarFunc);
+            node_fn_a.set_sig(ScalarFuncSig::CastIntAsInt);
+            node_fn_a
+                .mut_field_type()
+                .as_mut_accessor()
+                .set_tp(FieldTypeTp::Double);
+            node_fn_a.mut_children().push(node_fn_c);
+            node_fn_a
+        };
+
+        let node_fn_a_2 = {
+            // node e
+            let mut node_e = Expr::new();
+            node_e.set_tp(ExprType::Float64);
+            node_e
+                .mut_field_type()
+                .as_mut_accessor()
+                .set_tp(FieldTypeTp::Double);
+            node_e.mut_val().encode_f64(-1.5).unwrap();
+
+            // node f
+            let mut node_f = Expr::new();
+            node_f.set_tp(ExprType::Float64);
+            node_f
+                .mut_field_type()
+                .as_mut_accessor()
+                .set_tp(FieldTypeTp::Double);
+            node_f.mut_val().encode_f64(100.12).unwrap();
+
+            // FnB
+            let mut node_fn_b = Expr::new();
+            node_fn_b.set_tp(ExprType::ScalarFunc);
+            node_fn_b.set_sig(ScalarFuncSig::CastIntAsReal);
+            node_fn_b
+                .mut_field_type()
+                .as_mut_accessor()
+                .set_tp(FieldTypeTp::LongLong);
+            node_fn_b.mut_children().push(node_e);
+            node_fn_b.mut_children().push(node_f);
+
+            // FnA
+            let mut node_fn_a = Expr::new();
+            node_fn_a.set_tp(ExprType::ScalarFunc);
+            node_fn_a.set_sig(ScalarFuncSig::CastIntAsInt);
+            node_fn_a
+                .mut_field_type()
+                .as_mut_accessor()
+                .set_tp(FieldTypeTp::Double);
+            node_fn_a.mut_children().push(node_fn_b);
+            node_fn_a
+        };
+
+        // node a (NULL)
+        let mut node_a = Expr::new();
+        node_a.set_tp(ExprType::Null);
+        node_a
+            .mut_field_type()
+            .as_mut_accessor()
+            .set_tp(FieldTypeTp::Double);
+
+        // FnD
+        let mut node_fn_d = Expr::new();
+        node_fn_d.set_tp(ExprType::ScalarFunc);
+        node_fn_d.set_sig(ScalarFuncSig::CastIntAsDecimal);
+        node_fn_d
+            .mut_field_type()
+            .as_mut_accessor()
+            .set_tp(FieldTypeTp::Double);
+        node_fn_d.mut_children().push(node_a);
+        node_fn_d.mut_children().push(node_fn_a_1);
+        node_fn_d.mut_children().push(node_fn_a_2);
+
+        let mut vec = vec![];
+        append_rpn_nodes_recursively(node_fn_d, &mut vec, &Tz::utc(), fn_mapper).unwrap();
+
+        let mut it = vec.into_iter();
+
+        // node a
+        assert!(it
+            .next()
+            .unwrap()
+            .constant_value()
+            .unwrap()
+            .as_real()
+            .is_none());
+
+        // node b
+        assert_eq!(
+            it.next()
+                .unwrap()
+                .constant_value()
+                .unwrap()
+                .as_int()
+                .unwrap(),
+            7
+        );
+
+        // node c
+        assert_eq!(
+            it.next()
+                .unwrap()
+                .constant_value()
+                .unwrap()
+                .as_int()
+                .unwrap(),
+            3
+        );
+
+        // node d
+        assert_eq!(
+            it.next()
+                .unwrap()
+                .constant_value()
+                .unwrap()
+                .as_int()
+                .unwrap(),
+            11
+        );
+
+        // FnC
+        assert_eq!(it.next().unwrap().fn_call_func().unwrap().name(), "FnC");
+
+        // FnA
+        assert_eq!(it.next().unwrap().fn_call_func().unwrap().name(), "FnA");
+
+        // node e
+        assert_eq!(
+            it.next()
+                .unwrap()
+                .constant_value()
+                .unwrap()
+                .as_real()
+                .unwrap(),
+            -1.5
+        );
+
+        // node f
+        assert_eq!(
+            it.next()
+                .unwrap()
+                .constant_value()
+                .unwrap()
+                .as_real()
+                .unwrap(),
+            100.12
+        );
+
+        // FnB
+        assert_eq!(it.next().unwrap().fn_call_func().unwrap().name(), "FnB");
+
+        // FnA
+        assert_eq!(it.next().unwrap().fn_call_func().unwrap().name(), "FnA");
+
+        // FnD
+        assert_eq!(it.next().unwrap().fn_call_func().unwrap().name(), "FnD");
+
+        // Finish
+        assert!(it.next().is_none())
+    }
+}
